requirements.txt
python-dotenv-1.2.2
bcrypt-5.0.0
django-5.2.12
djangorestframework-3.16.1
psycopg2-binary-2.9.11
pyjwt-2.12.1

